+++ src/TextInputHelper.h
@@ -51,7 +51,7 @@ public:
     static void Reset();
 
     static string GetCurrentText() { return currentText; }
-    static void SetCurrentText(const string &currentText) { TextInputHelper::currentText = currentText; caretPosition = currentText.length(); }
+    static void SetCurrentText(const string &currentText) { TextInputHelper::currentText = currentText; caretPosition = (int)currentText.length(); }
     static unsigned int GetCaretPosition() { return caretPosition; }
     static bool GetIsCaretShowing() { return isCaretShowing; }
     static void SetMaxPixelWidth(int pixelWidth, MLIFont *pFontToCheckAgainst) { maxPixelWidth = pixelWidth; TextInputHelper::pFontToCheckAgainst = pFontToCheckAgainst; }
